requirements.txt
biopython
pandas
matplotlib
seaborn
openpyxl
requests
matplotlib-venn